+++ src/main/java/org/monarchinitiative/lr2pg/command/HpoCase2SvgCommand.java
@@ -9,6 +9,7 @@ import org.monarchinitiative.lr2pg.likelihoodratio.TestResult;
 import org.monarchinitiative.lr2pg.svg.Lr2Svg;
 import org.monarchinitiative.phenol.formats.hpo.HpoDisease;
 import org.monarchinitiative.phenol.formats.hpo.HpoOntology;
+import org.monarchinitiative.phenol.ontology.data.TermId;
 
 import java.io.IOException;
 
@@ -20,14 +21,15 @@ import java.io.IOException;
 public class HpoCase2SvgCommand implements Command {
     private static final Logger logger = LogManager.getLogger();
     private final String dataDirectory;
-
-    private final String diseaseName;
+    /** A disease CURIE such as OMIM:600100. */
+    private final TermId diseaseCurie;
 
     private final int n_terms_per_case;
     private final int n_noise_terms;
 
-    public HpoCase2SvgCommand(String datadir, String diseaseName, String outfileName, int terms_per_case, int noise_terms ) {
-        this.diseaseName=diseaseName;
+    public HpoCase2SvgCommand(String datadir, String diseaseId, String outfileName, int terms_per_case, int noise_terms ) {
+
+        this.diseaseCurie = TermId.constructWithPrefix(diseaseId);
         this.dataDirectory=datadir;
         this.n_terms_per_case=terms_per_case;
         this.n_noise_terms=noise_terms;
@@ -38,7 +40,7 @@ public class HpoCase2SvgCommand implements Command {
         int cases_to_simulate=1;
         HpoCaseSimulator simulator = new HpoCaseSimulator( dataDirectory, cases_to_simulate, n_terms_per_case, n_noise_terms);
         try {
-            HpoDisease disease = simulator.name2disease(diseaseName);
+            HpoDisease disease = simulator.name2disease(diseaseCurie);
             HpoOntology ontology = simulator.getOntology();
             simulator.simulateCase(disease);
             TestResult result = simulator.getResults(disease);