+++ src/main/java/org/monarchinitiative/lr2pg/genotype/Genotype2LR.java
@@ -1,0 +1,4 @@
+package org.monarchinitiative.lr2pg.genotype;
+
+public class Genotype2LR {
+}