+++ src/main/java/org/monarchinitiative/lr2pg/io/FileDownloader.java
@@ -117,9 +117,12 @@ public class FileDownloader {
             // Try to get file size.
             FTPFile[] files = ftp.listFiles(fileName);
             long fileSize = -1;
-            for (int i = 0; i < files.length; ++i)
-                if (files[i].getName().equals(fileName))
-                    fileSize = files[i].getSize();
+            for (FTPFile ftpfile : files) {
+                if (ftpfile.getName().equals(fileName)) {
+                    fileSize=ftpfile.getSize();
+                    break;
+                }
+            }
             ftp.pwd();
             ProgressBar pb = null;
             if (fileSize != -1)
@@ -233,7 +236,8 @@ public class FileDownloader {
             out.close();
             if (pb != null && pos != pb.getMax())
                 pb.print(fileSize);
-        } catch (IOException e) {
+        } catch (IOException | IllegalStateException e) {
+            logger.error(String.format("Failed to downloaded file from %s",src.getHost()),e);
             throw new FileDownloadException("ERROR: Problem downloading file: " + e.getMessage());
         }
         return true;