+++ src/main/java/org/monarchinitiative/lr2pg/likelihoodratio/GenotypeLikelihoodRatio.java
@@ -1,0 +1,9 @@
+package org.monarchinitiative.lr2pg.likelihoodratio;
+
+public class GenotypeLikelihoodRatio {
+
+
+
+
+
+}