+++ src/main/java/org/monarchinitiative/lr2pg/poisson/MaxCountExceededException.java
@@ -1,0 +1,8 @@
+package org.monarchinitiative.lr2pg.poisson;
+
+public class MaxCountExceededException extends Exception {
+
+    public MaxCountExceededException(int mc) {
+        super(String.valueOf(mc));
+    }
+}