+++ CHANGELOG.rst
@@ -3,6 +3,12 @@ Changelog
 =========
 
 ------
+v0.4.0
+------
+- refactored to use phenol 1.1.0
+- adding code to simulate genotypes
+
+------
 v0.3.2
 ------
 - refactored TermId to remove superfluous interface and renamed ImmutableTermId to TermId